A few minor updates
