+++ .github/workflows/package.yml
@@ -50,9 +50,8 @@ jobs:
           conda config --set solver libmamba
 
           conda activate hexrd
-          conda install --override-channels -c conda-forge anaconda-client conda-build conda pybind11
+          conda install --override-channels -c conda-forge anaconda-client conda-build conda
 
-          python install_build_dependencies.py
       # This is need to ensure ~/.profile or ~/.bashrc are used so the activate
       # command works.
       shell: bash -l {0}
